`AbstractReactiveSocket.onClose()` never completes.
#### Problem

`AbstractReactiveSocket` does not implement `close()` and `onClose()` correctly. Both of them return `Px.never()`.
This makes it hard for the implementation to do any cleanup actions on close.

#### Modification

Correctly terminate `onClose` `Publisher` after close() is invoked and subscribed.
Also, added test to verify this behavior in the local transport.

#### Result

Better way to cleanup on close of `AbstractReactiveSocket`.